+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Possible sections are:
 - `Resolution::area_rads2`
 - `Resolution::edge_length_rads`
 - `CellIndex::child_position`
+- `CellIndex::child_at`
 
 ### Changed